引数の型名不一致でエラーが出ていたのを修正。u32に。

+++ dxlibp/cpcnv/utf8/utf8toucs2.c
@@ -1,6 +1,6 @@
 #include "../../cpcnv.h"
 
-int dxpCpUTF8_toUcs2(dxpUcs2 * dst, size_t count,const dxpChar * src)
+int dxpCpUTF8_toUcs2(dxpUcs2 * dst, u32 count,const dxpChar * src)
 {
 	if (!src || !dst || !count) return 0;
